# Rerender optimisations