--- a/app/react-native/src/preview/components/StoryView/StoryView.tsx
+++ b/app/react-native/src/preview/components/StoryView/StoryView.tsx
@@ -39,4 +39,4 @@ const StoryView = ({ context }: Props) => {
   );
 };
 
-export default StoryView;
+export default React.memo(StoryView);