Issue:

A lot of interactions in Storybook are much less responsive than they should be, the profiler indicates that large tree hierarchies are rerendered that result in no UI changes.

## What I did

- Profile various interactions
  - Switching stories
  - Adjust addon properties (knobs, controls, background, etc.)
  - Switching navigation / preview / addons tabs
- Apply memoization to props that were frequent causes of rerenders, when their data hadn't changed meaningfully
- Apply memoization to components that can often avoid being rerendered (such as when switching tabs) 
- Introduced [jotai](https://jotai.org/) to lift the story context state out of `OnDeviceUI`'s parent, and read the state closer to where it needs to be instead of using prop-drilling

Below are screenshots from the profiler of the longest commit when switching stories. The gist of it is that views that were previously re-rendering just because the parent re-rendered are now not doing that, and only the relevant story list items (the one that was selected, and the one that will become selected) are updating.

The net result is a significant improvement to responsiveness, even in a dev build.

The exact numbers are somewhat unreliable because it's not a release build, but that factor between them is probably a reasonable indication.

| Before | After |
|------|---------------|
|  ![image](https://user-images.githubusercontent.com/112166/219486273-b8ad77d1-5462-4a27-a3a9-38e0386e038b.png)  | ![image](https://user-images.githubusercontent.com/112166/219486245-1dfaae19-62b1-440e-bc56-df15825e383e.png) |

## How to test

Unfortunately there are no benchmark/performance tests, so it has to mostly be manually by performing typical Storybook operations:

- Switching stories
- Switching tabs
- Filtering stories
- Manipulating addon controls

<!--

Everybody: Please submit all PRs to the `next-6.0` branch unless they are specific to 5.3. Storybook maintainers cherry-pick bug and documentation fixes into the `master` branch as part of the release process, so you shouldn't need to worry about this.

Maintainers: Please tag your pull request with at least one of the following:
`["cleanup", "BREAKING CHANGE", "feature request", "bug", "documentation", "maintenance", "dependencies", "other"]`

-->